Splitted dependencywheel series into several files.
