+++ js/Series/DependencyWheel/DependencyWheelSeries.js
@@ -27,10 +27,16 @@ import A from '../../Core/Animation/AnimationUtilities.js';
 var animObject = A.animObject;
 import BaseSeries from '../../Core/Series/Series.js';
 var _a = BaseSeries.seriesTypes, PieSeries = _a.pie, SankeySeries = _a.sankey;
+import DependencyWheelPoint from './DependencyWheelPoint.js';
 import H from '../../Core/Globals.js';
-import NodesMixin from '../../Mixins/Nodes.js';
+var deg2rad = H.deg2rad;
 import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
+/* *
+ *
+ *  Class
+ *
+ * */
 /**
  * @private
  * @class
@@ -168,7 +174,7 @@ var DependencyWheelSeries = /** @class */ (function (_super) {
      */
     DependencyWheelSeries.prototype.translate = function () {
         var options = this.options, factor = 2 * Math.PI /
-            (this.chart.plotHeight + this.getNodePadding()), center = this.getCenter(), startAngle = (options.startAngle - 90) * H.deg2rad;
+            (this.chart.plotHeight + this.getNodePadding()), center = this.getCenter(), startAngle = (options.startAngle - 90) * deg2rad;
         SankeySeries.prototype.translate.call(this);
         this.nodeColumns[0].forEach(function (node) {
             // Don't render the nodes if sum is 0 #12453
@@ -288,74 +294,6 @@ var DependencyWheelSeries = /** @class */ (function (_super) {
 extend(DependencyWheelSeries.prototype, {
     orderNodes: false,
     getCenter: PieSeries.prototype.getCenter
-});
-/* *
- *
- *  Class
- *
- * */
-var DependencyWheelPoint = /** @class */ (function (_super) {
-    __extends(DependencyWheelPoint, _super);
-    function DependencyWheelPoint() {
-        /* *
-         *
-         *  Properties
-         *
-         * */
-        var _this = _super !== null && _super.apply(this, arguments) || this;
-        _this.angle = void 0;
-        _this.fromNode = void 0;
-        _this.index = void 0;
-        _this.linksFrom = void 0;
-        _this.linksTo = void 0;
-        _this.options = void 0;
-        _this.series = void 0;
-        _this.shapeArgs = void 0;
-        _this.toNode = void 0;
-        return _this;
-        /* eslint-enable valid-jsdoc */
-    }
-    /* *
-     *
-     *  Functions
-     *
-     * */
-    /* eslint-disable valid-jsdoc */
-    /**
-     * Return a text path that the data label uses.
-     * @private
-     */
-    DependencyWheelPoint.prototype.getDataLabelPath = function (label) {
-        var renderer = this.series.chart.renderer, shapeArgs = this.shapeArgs, upperHalf = this.angle < 0 || this.angle > Math.PI, start = shapeArgs.start, end = shapeArgs.end;
-        if (!this.dataLabelPath) {
-            this.dataLabelPath = renderer
-                .arc({
-                open: true,
-                longArc: Math.abs(Math.abs(start) - Math.abs(end)) < Math.PI ? 0 : 1
-            })
-                // Add it inside the data label group so it gets destroyed
-                // with the label
-                .add(label);
-        }
-        this.dataLabelPath.attr({
-            x: shapeArgs.x,
-            y: shapeArgs.y,
-            r: (shapeArgs.r +
-                (this.dataLabel.options.distance || 0)),
-            start: (upperHalf ? start : end),
-            end: (upperHalf ? end : start),
-            clockwise: +upperHalf
-        });
-        return this.dataLabelPath;
-    };
-    DependencyWheelPoint.prototype.isValid = function () {
-        // No null points here
-        return true;
-    };
-    return DependencyWheelPoint;
-}(SankeySeries.prototype.pointClass));
-extend(DependencyWheelPoint.prototype, {
-    setState: NodesMixin.setNodeState
 });
 DependencyWheelSeries.prototype.pointClass = DependencyWheelPoint;
 BaseSeries.registerSeriesType('dependencywheel', DependencyWheelSeries);
