+++ ts/Series/DependencyWheel/DependencyWheelPointOptions.d.ts
@@ -1,0 +1,31 @@
+/* *
+ *
+ *  Dependency wheel module
+ *
+ *  (c) 2018-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type SankeyPointOptions from '../Sankey/SankeyPointOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface DependencyWheelPointOptions extends SankeyPointOptions {
+    // nothing here yet
+}
+
+export default DependencyWheelPointOptions;
